requirements.txt
mock
lxml
pyOpenSSL
requests
urllib3